services/agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "review-agent"
version = "0.1.0"
description = "FastAPI agent service for the review system"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "langgraph>=0.2,<1.0",
  "openai>=1.0,<2.0",
  "python-dotenv>=1.0,<2.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
test = [
  "httpx>=0.28,<1.0",
  "pytest>=8.0,<9.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
